refactor: changes the logic of the points to show decimal digit

+++ guess.c
@@ -11,7 +11,7 @@ int main() {
 
   int guess;
   int numberOfAttempts = 1;
-  int losing_points = 0;
+  double losing_points = 0;
 
   while(1) {
     printf("\n****** Tentativa n.º%d ******\n\n", numberOfAttempts);
@@ -33,7 +33,7 @@ int main() {
 
       break;
     } else {
-      losing_points += (guess - secretNumber) / 2;
+      losing_points += (guess - secretNumber) / 2.0;
 
       if (guess > secretNumber) {
         printf("Seu chute foi maior que o número secreto!\n");
@@ -44,10 +44,9 @@ int main() {
 
     numberOfAttempts++;
   }
-
-  int totalPoints = INITIAL_POINTS - losing_points;
+  double totalPoints = INITIAL_POINTS - losing_points;
 
   printf("\nForam realizadas %d tentativas.\n", numberOfAttempts);
-  printf("Total de pontos: %d", totalPoints);
+  printf("Total de pontos: %.1f", totalPoints);
   printf("\n****** Fim de game! ******");
 }
